Begining refactor to fix #77.
